TriggerSchema can checkTrigger
