+++ projects/Eevee/souchy.celebi.eevee/impl/shared/triggers/schemas/TriggerOnMove.cs
@@ -1,4 +1,5 @@
 ﻿using souchy.celebi.eevee.enums;
+using souchy.celebi.eevee.face.objects;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -7,12 +8,17 @@ using System.Threading.Tasks;
 
 namespace souchy.celebi.eevee.impl.shared.triggers.schemas
 {
-    internal class TriggerOnMove : TriggerSchema
+    public class TriggerOnMove : TriggerSchema
     {
         /// <summary>
         /// Walk, Teleport, Swap, Translate...
         /// </summary>
         public MoveType moveType { get; set; }
+
+        public override bool checkTrigger(IAction action, TriggerEvent triggerEvent)
+        {
+            return true;
+        }
     }
 
     public class TriggerOnCellMovement : TriggerSchema
@@ -23,5 +29,10 @@ namespace souchy.celebi.eevee.impl.shared.triggers.schemas
         /// stop on cell
         /// </summary>
         public CellMovementType moveType { get; set; }
+
+        public override bool checkTrigger(IAction action, TriggerEvent triggerEvent)
+        {
+            return true;
+        }
     }
 }
