+++ projects/Eevee/souchy.celebi.eevee/impl/shared/triggers/schemas/TriggerOnTimeline.cs
@@ -1,4 +1,5 @@
 ﻿using souchy.celebi.eevee.enums;
+using souchy.celebi.eevee.face.objects;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -14,13 +15,20 @@ namespace souchy.celebi.eevee.impl.shared.triggers.schemas
 
     public class TriggerOnTimeline : TriggerSchema
     {
-        public MomentType moment { get; set; } = MomentType.TurnStart;
         // fight start
         // fight end
         // round start
         // round end
         // turn start
         // turn end
+        public MomentType moment { get; set; } = MomentType.TurnStart;
+
+        public override bool checkTrigger(IAction action, TriggerEvent triggerEvent)
+        {
+            TriggerEventTimeline timelineEvent = (TriggerEventTimeline) triggerEvent;
+
+            return timelineEvent.moment.Equals(moment);
+        }
     }
 
 }
